Use an array of longs not guint32s for a property change with format 32.
2003-05-26  Mark McLoughlin  <mark@skynet.ie>

        * panel-xutils.c: (panel_xutils_set_strut): Use an array of
        longs not guint32s for a property change with format 32. Doh.
        Might well fix #113475.

+++ gnome-panel/panel-xutils.c
@@ -92,7 +92,7 @@ panel_xutils_set_strut (GdkWindow *gdk_window,
 {
 	Display *display;
 	Window   window;
-	guint32  struts [4];
+	gulong   struts [4];
 
 	g_return_if_fail (GDK_IS_WINDOW (gdk_window));
